More trait fixes and testing cleanup.
- Removed `AlBackendName` from op_dispatcher.h that I forgot to remove
and updated uses.
- Moved some algorithm description into a separate file to cut down on
circular includes.
