+++ test/test_utils_ht.hpp
@@ -84,44 +84,3 @@ void complete_operations<Al::HostTransferBackend>(
   typename Al::HostTransferBackend::comm_type& comm) {
   AL_FORCE_CHECK_GPU_NOSYNC(AlGpuStreamSynchronize(comm.get_stream()));
 }
-
-// Host-transfer allreduce supports two algorithms.
-template <>
-std::vector<std::pair<std::string, typename Al::OpAlgoType<Al::AlOperation::allreduce, Al::HostTransferBackend>::type>> get_supported_algos<Al::AlOperation::allreduce, Al::HostTransferBackend>() {
-  using algo_type = Al::HostTransferBackend::allreduce_algo_type;
-  return {{"automatic", algo_type::automatic},
-          {"host_transfer", algo_type::host_transfer}};
-}
-
-// Define default algorithms for each operation.
-template <>
-struct AlgorithmOptions<Al::HostTransferBackend> {
-  typename Al::HostTransferBackend::allgather_algo_type allgather_algo =
-    Al::HostTransferBackend::allgather_algo_type::automatic;
-  typename Al::HostTransferBackend::allgatherv_algo_type allgatherv_algo =
-    Al::HostTransferBackend::allgatherv_algo_type::automatic;
-  typename Al::HostTransferBackend::allreduce_algo_type allreduce_algo =
-    Al::HostTransferBackend::allreduce_algo_type::automatic;
-  typename Al::HostTransferBackend::alltoall_algo_type alltoall_algo =
-    Al::HostTransferBackend::alltoall_algo_type::automatic;
-  typename Al::HostTransferBackend::alltoallv_algo_type alltoallv_algo =
-    Al::HostTransferBackend::alltoallv_algo_type::automatic;
-  typename Al::HostTransferBackend::barrier_algo_type barrier_algo =
-    Al::HostTransferBackend::barrier_algo_type::automatic;
-  typename Al::HostTransferBackend::bcast_algo_type bcast_algo =
-    Al::HostTransferBackend::bcast_algo_type::automatic;
-  typename Al::HostTransferBackend::gather_algo_type gather_algo =
-    Al::HostTransferBackend::gather_algo_type::automatic;
-  typename Al::HostTransferBackend::gatherv_algo_type gatherv_algo =
-    Al::HostTransferBackend::gatherv_algo_type::automatic;
-  typename Al::HostTransferBackend::reduce_algo_type reduce_algo =
-    Al::HostTransferBackend::reduce_algo_type::automatic;
-  typename Al::HostTransferBackend::reduce_scatter_algo_type reduce_scatter_algo =
-    Al::HostTransferBackend::reduce_scatter_algo_type::automatic;
-  typename Al::HostTransferBackend::reduce_scatterv_algo_type reduce_scatterv_algo =
-    Al::HostTransferBackend::reduce_scatterv_algo_type::automatic;
-  typename Al::HostTransferBackend::scatter_algo_type scatter_algo =
-    Al::HostTransferBackend::scatter_algo_type::automatic;
-  typename Al::HostTransferBackend::scatterv_algo_type scatterv_algo =
-    Al::HostTransferBackend::scatterv_algo_type::automatic;
-};
