+++ test/test_utils_mpi.hpp
@@ -26,42 +26,3 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
-
-#include "Al.hpp"
-#include "aluminum/traits/traits.hpp"
-
-#include "test_utils.hpp"
-
-
-// Define default algorithms for each operation.
-template <>
-struct AlgorithmOptions<Al::MPIBackend> {
-  typename Al::MPIBackend::allgather_algo_type allgather_algo =
-    Al::MPIBackend::allgather_algo_type::automatic;
-  typename Al::MPIBackend::allgatherv_algo_type allgatherv_algo =
-    Al::MPIBackend::allgatherv_algo_type::automatic;
-  typename Al::MPIBackend::allreduce_algo_type allreduce_algo =
-    Al::MPIBackend::allreduce_algo_type::automatic;
-  typename Al::MPIBackend::alltoall_algo_type alltoall_algo =
-    Al::MPIBackend::alltoall_algo_type::automatic;
-  typename Al::MPIBackend::alltoallv_algo_type alltoallv_algo =
-    Al::MPIBackend::alltoallv_algo_type::automatic;
-  typename Al::MPIBackend::barrier_algo_type barrier_algo =
-    Al::MPIBackend::barrier_algo_type::automatic;
-  typename Al::MPIBackend::bcast_algo_type bcast_algo =
-    Al::MPIBackend::bcast_algo_type::automatic;
-  typename Al::MPIBackend::gather_algo_type gather_algo =
-    Al::MPIBackend::gather_algo_type::automatic;
-  typename Al::MPIBackend::gatherv_algo_type gatherv_algo =
-    Al::MPIBackend::gatherv_algo_type::automatic;
-  typename Al::MPIBackend::reduce_algo_type reduce_algo =
-    Al::MPIBackend::reduce_algo_type::automatic;
-  typename Al::MPIBackend::reduce_scatter_algo_type reduce_scatter_algo =
-    Al::MPIBackend::reduce_scatter_algo_type::automatic;
-  typename Al::MPIBackend::reduce_scatterv_algo_type reduce_scatterv_algo =
-    Al::MPIBackend::reduce_scatterv_algo_type::automatic;
-  typename Al::MPIBackend::scatter_algo_type scatter_algo =
-    Al::MPIBackend::scatter_algo_type::automatic;
-  typename Al::MPIBackend::scatterv_algo_type scatterv_algo =
-    Al::MPIBackend::scatterv_algo_type::automatic;
-};
